Move FileService and GitService to services module and update imports
